Created processing to zombie
