+++ ZombieAI/src/main/java/zombieaisystem/AIProcessor.java
@@ -4,9 +4,7 @@ import common.data.Entity;
 import common.data.GameData;
 import common.data.World;
 import common.data.entities.player.Player;
-import common.data.entities.zombie.IZombieAI;
 import common.data.entities.zombie.Zombie;
-import common.data.entityparts.MovingPart;
 import common.data.entityparts.PositionPart;
 import common.services.IPostEntityProcessingService;
 
@@ -15,7 +13,7 @@ import java.util.List;
 
 import static common.data.GameData.TILE_SIZE;
 
-public class AIProcessor implements IPostEntityProcessingService, IZombieAI {
+public class AIProcessor implements IPostEntityProcessingService {
     private AStar aStar = new AStar();
     private float AITime = 0.0F;
     @Override
@@ -49,68 +47,5 @@ public class AIProcessor implements IPostEntityProcessingService, IZombieAI {
             String GOAL_STATE = (int)(playerPosition.getX() / TILE_SIZE) + "," + (int)(playerPosition.getY() / TILE_SIZE);
             zombie.setPathFinding(aStar.treeSearch(map, INITIAL_STATE, GOAL_STATE));
         }
-    }
-    @Override
-    public void moveTowards(GameData gameData, Entity zombie) {
-        PositionPart zombiePosition = zombie.getPart(PositionPart.class);
-        MovingPart zombieMovement = zombie.getPart(MovingPart.class);
-
-        List<int[]> pathFinding = ((Zombie) zombie).getPathFinding();
-        if (pathFinding == null) {
-            return;
-        }
-
-        float currentX = (int) zombiePosition.getX();
-        float currentY = (int) zombiePosition.getY();
-
-        int targetX;
-        int targetY;
-
-        if (pathFinding.size() <= 2) {
-            targetX = pathFinding.get(pathFinding.size() - 1)[0] * TILE_SIZE;
-            targetY = pathFinding.get(pathFinding.size() - 1)[1] * TILE_SIZE;
-        } else {
-            targetX = pathFinding.get(2)[0] * TILE_SIZE;
-            targetY = pathFinding.get(2)[1] * TILE_SIZE;
-
-            if (targetX == currentX && targetY == currentY) {
-                pathFinding.remove(2);
-                if (pathFinding.size() >= 3) {
-                    targetX = pathFinding.get(2)[0] * TILE_SIZE;
-                    targetY = pathFinding.get(2)[1] * TILE_SIZE;
-                }
-            }
-        }
-
-        float diffX = targetX - currentX;
-        float diffY = targetY - currentY;
-        float delta = gameData.getDelta();
-
-        if (Math.abs(diffX) <= Math.abs(zombieMovement.getDx() * delta)) {
-            zombieMovement.setDx(0);
-            zombiePosition.setX(currentX);
-        } else {
-            zombieMovement.setRight(diffX > 0);
-            zombieMovement.setLeft(diffX < 0);
-        }
-        if (Math.abs(diffY) <= Math.abs(zombieMovement.getDy() * delta)) {
-            zombieMovement.setDy(0);
-            zombiePosition.setY(currentY);
-        } else {
-            zombieMovement.setUp(diffY > 0);
-            zombieMovement.setDown(diffY < 0);
-        }
-
-        zombieMovement.process(gameData, zombie);
-        zombiePosition.process(gameData, zombie);
-
-        zombieMovement.setUp(false);
-        zombieMovement.setDown(false);
-        zombieMovement.setRight(false);
-        zombieMovement.setLeft(false);
-    }
-
-    public void setAITime(float AITime) {
-        this.AITime = AITime;
     }
 }
